+++ src/utils/listener_windows.c
@@ -43,7 +43,7 @@ static LRESULT CALLBACK ClipWindProc(HWND window, UINT msg, WPARAM wParam, LPARA
             if (check_and_delete_temp_file()) {  // send text only if it's not from clip-share
                 return 0;
             }
-            int copied_type = get_copied_type();
+            int8_t copied_type = get_copied_type();
             if (copied_type != COPIED_TYPE_NONE) {
                 clip_callback(copied_type);
             }
